# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Proof Fabric Protocol (PFP) API
  description: Cryptographically verifiable Financial Evidence Artifacts — enterprise
    control plane.
  version: 2.0.0
paths:
  /api/fea/generate:
    post:
      tags:
      - FEA
      summary: Generate Fea Endpoint
      description: Generate a Financial Evidence Artifact (tenant-bound, dual-layer
        replay protection).
      operationId: generate_fea_endpoint_api_fea_generate_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateFEARequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FEAResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /api/fea/batch:
    post:
      tags:
      - FEA
      summary: Batch Generate Fea
      description: 'Generate up to 500 FEAs in a single call. Each item is processed

        independently; partial failures are reported per-item.'
      operationId: batch_generate_fea_api_fea_batch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGenerateRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGenerateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /api/fea/verify:
    post:
      tags:
      - FEA
      summary: Verify Fea Endpoint
      description: Verify an FEA payload and signature.
      operationId: verify_fea_endpoint_api_fea_verify_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyFEARequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyFEAResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /api/fea:
    get:
      tags:
      - FEA
      summary: List Feas
      description: List the authenticated tenant's FEAs (paginated).
      operationId: list_feas_api_fea_get
      security:
      - APIKeyHeader: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: skip
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Skip
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FEAListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/fea/{fea_id}:
    get:
      tags:
      - FEA
      summary: Get Fea
      description: Authenticated single-FEA fetch (tenant scoped).
      operationId: get_fea_api_fea__fea_id__get
      security:
      - APIKeyHeader: []
      parameters:
      - name: fea_id
        in: path
        required: true
        schema:
          type: string
          title: Fea Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FEAResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/webhooks/subscribe:
    post:
      tags:
      - Webhooks
      summary: Subscribe
      operationId: subscribe_api_webhooks_subscribe_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - APIKeyHeader: []
  /api/webhooks:
    get:
      tags:
      - Webhooks
      summary: List Subscriptions
      operationId: list_subscriptions_api_webhooks_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - APIKeyHeader: []
  /api/webhooks/test:
    post:
      tags:
      - Webhooks
      summary: Test Webhook
      operationId: test_webhook_api_webhooks_test_post
      security:
      - APIKeyHeader: []
      parameters:
      - name: webhook_id
        in: query
        required: true
        schema:
          type: string
          title: Webhook Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/webhooks/{webhook_id}:
    delete:
      tags:
      - Webhooks
      summary: Delete Webhook
      operationId: delete_webhook_api_webhooks__webhook_id__delete
      security:
      - APIKeyHeader: []
      parameters:
      - name: webhook_id
        in: path
        required: true
        schema:
          type: string
          title: Webhook Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BatchGenerateRequest:
      properties:
        items:
          items:
            $ref: '#/components/schemas/GenerateFEARequest'
          type: array
          maxItems: 500
          minItems: 1
          title: Items
      type: object
      required:
      - items
      title: BatchGenerateRequest
    BatchGenerateResponse:
      properties:
        total:
          type: integer
          title: Total
        succeeded:
          type: integer
          title: Succeeded
        failed:
          type: integer
          title: Failed
        results:
          items:
            $ref: '#/components/schemas/BatchItemResult'
          type: array
          title: Results
      type: object
      required:
      - total
      - succeeded
      - failed
      - results
      title: BatchGenerateResponse
    BatchItemResult:
      properties:
        index:
          type: integer
          title: Index
        success:
          type: boolean
          title: Success
        fea_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Fea Id
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - index
      - success
      title: BatchItemResult
...
    FEAListItem:
      properties:
        fea_id:
          type: string
          title: Fea Id
        transaction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Transaction Id
        created_at:
          type: string
          title: Created At
        public_key_id:
          type: string
          title: Public Key Id
      type: object
      required:
      - fea_id
      - created_at
      - public_key_id
      title: FEAListItem
    FEAListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FEAListItem'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        skip:
          type: integer
          title: Skip
      type: object
      required:
      - items
      - total
      - limit
      - skip
      title: FEAListResponse
    FEAResponse:
      properties:
        fea_id:
          type: string
          title: Fea Id
        fea_payload:
          additionalProperties: true
          type: object
          title: Fea Payload
        signature:
          type: string
          title: Signature
        signature_version:
          type: string
          title: Signature Version
          default: v2
        public_key_id:
          type: string
          title: Public Key Id
        created_at:
          type: string
          title: Created At
      type: object
      required:
      - fea_payload
      - signature
      - public_key_id
      title: FEAResponse
      description: 'Response from FEA generation.


        Structure separates:

        - fea_payload: the signed data

        - signature: cryptographic signature

        - signature_version: signature metadata (NOT signed)

        - created_at: response metadata (NOT signed)'
    GenerateFEARequest:
      properties:
        idempotency_key:
          type: string
          minLength: 1
          title: Idempotency Key
        transaction_id:
          type: string
          minLength: 1
          title: Transaction Id
        timestamp:
          type: string
          title: Timestamp
          description: ISO 8601 UTC timestamp
        amount:
          type: integer
          minimum: 0.0
          title: Amount
          description: Amount in smallest unit (e.g., paise)
        currency:
          type: string
          maxLength: 3
          minLength: 3
          title: Currency
        payer_id:
          type: string
          title: Payer Id
          description: Hashed/tokenized payer identifier
        payee_id:
          type: string
          title: Payee Id
          description: Hashed/tokenized payee identifier
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
      type: object
      required:
      - idempotency_key
      - transaction_id
      - timestamp
      - amount
      - currency
      - payer_id
      - payee_id
      title: GenerateFEARequest
      description: Input for FEA generation - POST /generate-fea.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SubscribeRequest:
      properties:
        url:
          type: string
          maxLength: 2083
          minLength: 1
          format: uri
          title: Url
        events:
          items:
            type: string
          type: array
          title: Events
      type: object
      required:
      - url
      title: SubscribeRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    VerifyFEARequest:
      properties:
        fea_payload:
          additionalProperties: true
          type: object
          title: Fea Payload
        signature:
          type: string
          title: Signature
        signature_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Signature Version
      type: object
      required:
      - fea_payload
      - signature
      title: VerifyFEARequest
      description: 'Input for FEA verification - POST /verify-fea.


        Supports both:

        - New format: fea_payload + signature + signature_version (external)

        - Legacy format: fea_payload with signature_version inside + signature'
    VerifyFEAResponse:
      properties:
        valid:
          type: boolean
          title: Valid
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        signature_version:
          type: string
          title: Signature Version
          default: unknown
        verified_at:
          type: string
          title: Verified At
      type: object
      required:
      - valid
      title: VerifyFEAResponse
      description: Response from FEA verification.
  securitySchemes:
    APIKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://pfp-evidence.preview.emergentagent.com
  description: Sandbox/preview
- url: https://api.pfprotocol.com
  description: Production